Separate SiP logic in Redux code (#6945)
